refactor daemon bootstrap routes
